refactor!: Moved pomodoro.tsx
Moved /src/pages/Panel/pomodoro.tsx to /src/pages/Panel/components/pomodoro.tsx
